Fix integer overflow and null dereference
Resolves: COMPMID-4527

Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/5640
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>

+++ src/core/cpu/kernels/CpuMulKernel.cpp
@@ -846,7 +846,7 @@ void mul_S32_S32_S32(const ITensor *src1, const ITensor *src2, ITensor *out, con
                 }
                 else
                 {
-                    uint64_t mask = (1u << n) - 1;
+                    uint64_t mask = ((uint64_t)1u << n) - 1;
                     tmp           = (tmp + static_cast<int64_t>(mask)) >> n;
                 }
                 if(is_sat)
@@ -909,7 +909,7 @@ void mul_S32_S32_S32(const ITensor *src1, const ITensor *src2, ITensor *out, con
                 }
                 else
                 {
-                    uint64_t mask = (1u << n) - 1;
+                    uint64_t mask = ((uint64_t)1u << n) - 1;
                     tmp           = (tmp + static_cast<int64_t>(mask)) >> n;
                 }
                 if(is_sat)
